+++ src/icrar/leap-accelerate-cli/main.cc
@@ -22,7 +22,7 @@
 
 #include <icrar/leap-accelerate/utils.h>
 #include <icrar/leap-accelerate/math/Integration.h>
-#include <icrar/leap-accelerate/algorithm/cpu/PhaseRotate.h>
+#include <icrar/leap-accelerate/algorithm/casa/PhaseRotate.h>
 
 #include <casacore/measures/Measures/MDirection.h>
 
@@ -146,5 +146,5 @@ int main(int argc, char** argv)
     std::vector<casacore::MVDirection> directions; //ZenithDirection(ms);
     auto queue = std::queue<Integration>();
 
-    icrar::cpu::RemoteCalibration(*metadata, directions);
+    icrar::casa::RemoteCalibration(*metadata, directions);
 }
