+++ src/icrar/leap-accelerate/MetaData.cc
@@ -104,13 +104,13 @@ namespace icrar
         //Start calculations
         casacore::Matrix<double> A1;
         casacore::Array<std::int32_t> I1;
-        std::tie(A1, I1) = icrar::cpu::PhaseMatrixFunction(a1, a2, 0);
-        casacore::Matrix<double> Ad1 = icrar::cpu::PseudoInverse(A1);
+        std::tie(A1, I1) = icrar::casa::PhaseMatrixFunction(a1, a2, 0);
+        casacore::Matrix<double> Ad1 = icrar::casa::PseudoInverse(A1);
 
         casacore::Matrix<double> A;
         casacore::Array<std::int32_t> I;
-        std::tie(A, I) = icrar::cpu::PhaseMatrixFunction(a1, a2, -1);
-        casacore::Matrix<double> Ad = icrar::cpu::PseudoInverse(A);
+        std::tie(A, I) = icrar::casa::PhaseMatrixFunction(a1, a2, -1);
+        casacore::Matrix<double> Ad = icrar::casa::PseudoInverse(A);
 
         this->A = A;
         this->Ad = Ad;
