+++ src/icrar/leap-accelerate/math/casa/matrix.h
@@ -28,12 +28,11 @@
 
 namespace icrar
 {
-namespace cpu
+namespace casa
 {
     template<typename T>
     void multiply(const casacore::Matrix<T>& a, const casacore::Matrix<T>& b, casacore::Matrix<T>& c)
     {
-        //TODO: convert to cuda
         Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> ea = ConvertMatrix(a);
         Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> eb = ConvertMatrix(b);
         Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> ec = ea * eb;
@@ -51,7 +50,6 @@ namespace cpu
     template<typename T>
     void multiply(const casacore::Matrix<T>& a, const casacore::Array<T>& b, casacore::Array<T>& c)
     {
-        //TODO: convert to cuda
         Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic> ea = ConvertMatrix(a);
         Eigen::Matrix<T, Eigen::Dynamic, 1> eb = ConvertVector(b);
         Eigen::Matrix<T, Eigen::Dynamic, 1> ec = ea * eb;
